+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.8] - 2026-07-21
+
+### Added
+- Added Vulkan graphic API function hooks to the plugin interface
+- Added `vk_graphic_api_plugin` example.
+- Added `plugin_vulkan.h` for Vulkan plugin development.
+
+### Changed
+- Update plugin path name to small letters - `gameplug/plugins`
+
 ## [1.0.7] - 2026-07-18
 
 ### Added